+++ src/actions/index.js
@@ -110,3 +110,7 @@ export {
   getTranslationLocator,
   linkTranslation,
 } from '@plone/volto/actions/translations/translations';
+export {
+  setBlocksClipboard,
+  resetBlocksClipboard,
+} from '@plone/volto/actions/blocksClipboard/blocksClipboard';