+++ src/components/manage/Blocks/Block/Edit.jsx
@@ -139,7 +139,13 @@ class Edit extends Component {
         {Block !== null ? (
           <div
             role="presentation"
-            onClick={() => this.props.onSelectBlock(this.props.id)}
+            onClick={(e) => {
+              const isMultipleSelection = e.shiftKey;
+              this.props.onSelectBlock(
+                this.props.id,
+                this.props.selected ? false : isMultipleSelection,
+              );
+            }}
             onKeyDown={
               !(blockHasOwnFocusManagement || disableNewBlocks)
                 ? (e) =>
@@ -151,7 +157,10 @@ class Edit extends Component {
                     )
                 : null
             }
-            className={cx(`block ${type}`, { selected: this.props.selected })}
+            className={cx(`block ${type}`, {
+              selected: this.props.selected || this.props.multiSelected,
+              multiSelected: this.props.multiSelected,
+            })}
             style={{ outline: 'none' }}
             ref={this.blockNode}
             // The tabIndex is required for the keyboard navigation