+++ src/components/manage/Form/Form.jsx
@@ -43,6 +43,7 @@ import {
 } from 'semantic-ui-react';
 import { v4 as uuid } from 'uuid';
 import { toast } from 'react-toastify';
+import { BlocksClipboard } from '@plone/volto/components';
 import { settings } from '~/config';
 
 /**
@@ -172,6 +173,7 @@ class Form extends Component {
         formData[blocksLayoutFieldname].items.length > 0
           ? formData[blocksLayoutFieldname].items[0]
           : null,
+      multiSelected: [],
       placeholderProps: {},
       isClient: false,
     };
@@ -369,11 +371,36 @@ class Form extends Component {
    * Select block handler
    * @method onSelectBlock
    * @param {string} id Id of the field
+   * @param {string} isMultipleSelection true if multiple blocks are selected
    * @returns {undefined}
    */
-  onSelectBlock(id) {
+  onSelectBlock(id, isMultipleSelection) {
+    let multiSelected = [];
+
+    if (isMultipleSelection) {
+      const blocksLayoutFieldname = getBlocksLayoutFieldname(
+        this.state.formData,
+      );
+
+      const blocks_layout = this.state.formData[blocksLayoutFieldname].items;
+
+      const anchor =
+        this.state.multiSelected.length > 0
+          ? blocks_layout.indexOf(this.state.multiSelected[0])
+          : blocks_layout.indexOf(this.state.selected);
+      const focus = blocks_layout.indexOf(id);
+
+      if (anchor === focus) {
+        multiSelected = [id];
+      } else if (focus > anchor) {
+        multiSelected = [...blocks_layout.slice(anchor, focus + 1)];
+      } else {
+        multiSelected = [...blocks_layout.slice(focus, anchor + 1)];
+      }
+    }
     this.setState({
       selected: id,
+      multiSelected,
     });
   }
 
@@ -549,9 +576,10 @@ class Form extends Component {
    * @method onFocusPreviousBlock
    * @param {string} currentBlock The id of the current block
    * @param {node} blockNode The id of the current block
+   * @param {boolean} isMultipleSelection true if multiple blocks selected
    * @returns {undefined}
    */
-  onFocusPreviousBlock(currentBlock, blockNode) {
+  onFocusPreviousBlock(currentBlock, blockNode, isMultipleSelection) {
     const blocksLayoutFieldname = getBlocksLayoutFieldname(this.state.formData);
     const currentIndex = this.state.formData[
       blocksLayoutFieldname
@@ -566,6 +594,7 @@ class Form extends Component {
 
     this.onSelectBlock(
       this.state.formData[blocksLayoutFieldname].items[newindex],
+      isMultipleSelection,
     );
   }
 
@@ -574,9 +603,10 @@ class Form extends Component {
    * @method onFocusNextBlock
    * @param {string} currentBlock The id of the current block
    * @param {node} blockNode The id of the current block
+   * @param {boolean} isMultipleSelection true if multiple blocks selected
    * @returns {undefined}
    */
-  onFocusNextBlock(currentBlock, blockNode) {
+  onFocusNextBlock(currentBlock, blockNode, isMultipleSelection) {
     const blocksLayoutFieldname = getBlocksLayoutFieldname(this.state.formData);
     const currentIndex = this.state.formData[
       blocksLayoutFieldname
@@ -595,6 +625,7 @@ class Form extends Component {
 
     this.onSelectBlock(
       this.state.formData[blocksLayoutFieldname].items[newindex],
+      isMultipleSelection,
     );
   }
 
@@ -619,12 +650,13 @@ class Form extends Component {
       disableArrowDown = false,
     } = {},
   ) {
+    const isMultipleSelection = e.shiftKey;
     if (e.key === 'ArrowUp' && !disableArrowUp) {
-      this.onFocusPreviousBlock(block, node);
+      this.onFocusPreviousBlock(block, node, isMultipleSelection);
       e.preventDefault();
     }
     if (e.key === 'ArrowDown' && !disableArrowDown) {
-      this.onFocusNextBlock(block, node);
+      this.onFocusNextBlock(block, node, isMultipleSelection);
       e.preventDefault();
     }
     if (e.key === 'Enter' && !disableEnter) {
@@ -781,6 +813,23 @@ class Form extends Component {
       // but draftJS don't like it much and the hydration gets messed up
       this.state.isClient && (
         <div className="ui container">
+          <BlocksClipboard
+            formData={this.state.formData}
+            selectedBlocks={this.state.multiSelected}
+            onSetSelectedBlocks={(blockIds) =>
+              this.setState({ multiSelected: blockIds })
+            }
+            onSelectBlock={this.onSelectBlock}
+            onChangeBlocks={(newBlockData) =>
+              this.setState({
+                formData: {
+                  ...formData,
+                  ...newBlockData,
+                },
+              })
+            }
+            selectedBlock={this.state.selected}
+          />
           <DragDropContext
             onDragEnd={this.onDragEnd}
             onDragStart={this.handleDragStart}
@@ -837,6 +886,9 @@ class Form extends Component {
                               pathname={this.props.pathname}
                               block={block}
                               selected={this.state.selected === block}
+                              multiSelected={this.state.multiSelected.includes(
+                                block,
+                              )}
                               manage={this.props.isAdminForm}
                             />
                           </div>