+++ src/config/index.js
@@ -73,6 +73,7 @@ let config = {
     expressMiddleware: [],
     defaultBlockType: 'text',
     verticalFormTabs: false,
+    persistentReducers: ['blocksClipboard'],
   },
   widgets: {
     ...widgetMapping,