+++ src/helpers/MessageLabels/MessageLabels.js
@@ -162,4 +162,16 @@ export const messages = defineMessages({
     id: 'Users and Groups',
     defaultMessage: 'Users and Groups',
   },
+  copyBlocks: {
+    id: 'Copy blocks',
+    defaultMesages: 'Copy blocks',
+  },
+  cutBlocks: {
+    id: 'Cut blocks',
+    defaultMesages: 'Cut blocks',
+  },
+  pasteBlocks: {
+    id: 'Paste blocks',
+    defaultMesages: 'Paste blocks',
+  },
 });